correct prototype

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,4 +1,5 @@
 #include "lists.h"
+#include <stddef.h>
 #include <unistd.h>
 
 
@@ -51,7 +52,7 @@ void print_str(char *s)
  * Return: number of nodes printed
  */
 
-int print_list(list_t *node)
+size_t print_list(list_t *node)
 {
 	int i = 0;
